requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
pymongo>=4.4.0
python-dotenv>=1.0.0
colorama>=0.4.6
aiohttp>=3.8.0
asyncio>=3.4.3